Topic routes testing

+++ server/app/routes/topic.js
@@ -6,7 +6,6 @@ var Topic = db.model('topic');
 var Category = db.model('category');
 var Prerequisite = db.model('PrerequisiteTopic');
 
-
 module.exports = router;
 
 router.param('topicId', function(req, res, next, id) {
@@ -18,15 +17,21 @@ router.param('topicId', function(req, res, next, id) {
   }).catch(next)
 })
 
+// Fetches all topics
+// -- ordinary users: only approved topics
+// -- admins: all topics
 router.get('/', function(req, res, next) {
-  Topic.findAll()
+  var user = req.user,
+      whereCondition = { status: 'Approved' };
+  if(req.user && req.user.isAdmin) whereCondition = {}; // show all
+  Topic.findAll({ where: whereCondition })
   .then(topics => res.send(topics))
   .catch(next);
 });
 
 router.post('/', function(req, res, next) {
   Topic.create(req.body)
-  .then(topic => res.send(topic))
+  .then(topic => res.status(201).send(topic))
   .catch(next);
 });
 
@@ -38,14 +43,11 @@ router.get('/:topicId', function(req, res) {
 // ============================== ADMIN ROUTES ==============================
 
 router.put('/:topicId', function(req, res, next) {
-  if(req.user.isAdmin){
+  if(req.user && req.user.isAdmin){
     req.topic.update(req.body)
-    .then(topic => {
-      res.status(200).json(topic);
-    })
+    .then(topic => res.status(200).json(topic))
     .catch(next);
-}
-  else {
+  } else {
     var err = new Error('You must be an admin to update a topic');
     err.status = 401;
     throw err;
@@ -53,15 +55,15 @@ router.put('/:topicId', function(req, res, next) {
 });
 
 router.delete('/:topicId', function(req, res, next) {
-  if(req.user.isAdmin){
+  if(req.user && req.user.isAdmin){
     req.topic.destroy()
     .then(function(){
+      // we should replace these with model hooks to delete associations when a given topic is deleted
       return Promise.all([Category.destroy({where:{ topicId: req.topic.id}}), Prerequisite.destroy({where:{ topicId: req.topic.id} })]);
     })
-    .then(() => res.sendStatus(204))
+    .then(() => res.status(200).end())
     .catch(next);
-  }
-  else{
+  } else {
     var err = new Error('You must be an admin to delete a topic');
     err.status = 401;
     throw err;
